tests checks hernoemen en banklevels aan
bankleveltest werkt nog niet maar eerste opzet.
