+++ deprecated/tasks/tests_tasks/sqlite/profiles_used_task.py
@@ -3,10 +3,10 @@ from PyQt5.QtCore import pyqtSignal
 from qgis.core import QgsTask, Qgis
 from qgis.utils import QgsMessageLog, iface
 from hhnk_threedi_plugin.qgis_interaction.layers_management.adding_layers import add_layers
-from hhnk_threedi_plugin.gui.tests.sqlite_test_widgets.profiles_used_result import (
+from hhnk_threedi_plugin.gui.checks.sqlite_test_widgets.profiles_used_result import (
     create_profiles_used_widget,
 )
-from hhnk_threedi_plugin.gui.tests.sqlite_test_widgets.impervious_surface_result import (
+from hhnk_threedi_plugin.gui.checks.sqlite_test_widgets.impervious_surface_result import (
     create_impervious_surface_widget,
 )
