+++ hhnk_threedi_plugin/gui/checks/sqlite_test_widgets/isolated_channels_result.py
