+++ hhnk_threedi_plugin/gui/checks/sqlite_test_widgets/structs_channels_result.py
