+++ hhnk_threedi_plugin/hhnk_toolbox.py
@@ -45,12 +45,12 @@ from hhnk_threedi_plugin.hhnk_toolbox_dockwidget import HHNK_toolboxDockWidget
 from hhnk_threedi_plugin.gui.load_layers_popup import loadLayersDialog
 # from hhnk_threedi_plugin.gui.schematisation_splitter_uploader_dialog import schematisationDialog
 
-from hhnk_threedi_plugin.gui.tests.sqlite_check_popup import sqliteCheckDialog
-from hhnk_threedi_plugin.gui.tests.zero_d_one_d import zeroDOneDWidget
-from hhnk_threedi_plugin.gui.tests.one_d_two_d import oneDTwoDWidget
+from hhnk_threedi_plugin.gui.checks.sqlite_check_popup import sqliteCheckDialog
+from hhnk_threedi_plugin.gui.checks.zero_d_one_d import zeroDOneDWidget
+from hhnk_threedi_plugin.gui.checks.one_d_two_d import oneDTwoDWidget
 #from hhnk_threedi_plugin.gui.model_states.model_states import modelStateDialog
-from hhnk_threedi_plugin.gui.tests.sqlite_test_widgets.main_result_widget import collapsibleTree
-#from hhnk_threedi_plugin.gui.tests.bank_levels import bankLevelsWidget
+from hhnk_threedi_plugin.gui.checks.sqlite_test_widgets.main_result_widget import collapsibleTree
+from hhnk_threedi_plugin.gui.checks.bank_levels import bankLevelsWidget
 from hhnk_threedi_plugin.gui.klimaatsommen.klimaatsommen import KlimaatSommenWidget
 from hhnk_threedi_plugin.qgis_interaction.project import Project
 from hhnk_threedi_plugin.gui.new_project_dialog import newProjectDialog
@@ -493,7 +493,7 @@ class HHNK_toolbox:
                 self.input_data_dialog = inputDataDialog(caller=self, parent=self.dockwidget)
 
                 self.zero_d_one_d = zeroDOneDWidget(caller=self, parent=self.dockwidget)
-                #self.bank_levels = bankLevelsWidget(caller=self, parent=self.dockwidget)
+                self.bank_levels = bankLevelsWidget(caller=self, parent=self.dockwidget)
                 self.one_d_two_d = oneDTwoDWidget(caller=self, parent=self.dockwidget)
                 self.klimaatsommen = KlimaatSommenWidget(caller=self, parent=self.dockwidget)
                 self.notebook_widget = NotebookWidget(caller=self, parent=self.dockwidget)
